requirements.txt
transformers>=4.30.0
torch>=2.0.0
datasets>=2.12.0
numpy>=1.24.0
tqdm>=4.65.0
tensorboard>=2.13.0
pathlib>=1.0.1
regex>=2023.5.5
trl>=0.7.2
accelerate>=0.21.0
llama-cpp-python>=0.2.11
